requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
boto3==1.26.109
botocore==1.29.109
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.2.35
cryptography==40.0.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.3.0
Django==4.2
django-autoslug==1.9.9
django-celery-beat==2.5.0
django-cors-headers==3.14.0
django-countries==7.5.1
django-filter==23.1
django-heroku==0.3.1
django-manifest-loader==1.0.0
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-sms==0.6.0
django-storages==1.13.2
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.11.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.6.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.5.0
pipenv==2023.3.20
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
tornado==6.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
virtualenv==20.21.0
virtualenv-clone==0.5.7
wcwidth==0.2.6
whitenoise==6.4.0
